[Rename 🏷️] LoginInput으로 폴더명 수정
